backend/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
google-generativeai>=0.3.0
django-cors-headers>=4.3.0